# data-mannagement/data-mannagement/api/rolemanage/v1/role_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.6.2
// - protoc             v5.29.3
// source: rolemanage/v1/role.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRoleManageAddRole = "/api.rolemanage.v1.RoleManage/AddRole"
const OperationRoleManageDeleteRole = "/api.rolemanage.v1.RoleManage/DeleteRole"
const OperationRoleManageEditRole = "/api.rolemanage.v1.RoleManage/EditRole"
const OperationRoleManageEditRoleDisabled = "/api.rolemanage.v1.RoleManage/EditRoleDisabled"
const OperationRoleManageRoleDetail = "/api.rolemanage.v1.RoleManage/RoleDetail"
const OperationRoleManageRoleList = "/api.rolemanage.v1.RoleManage/RoleList"

type RoleManageHTTPServer interface {
	// AddRole新增角色
	AddRole(context.Context, *AddRoleRequest) (*AddRoleReply, error)
	// DeleteRole 删除角色
	DeleteRole(context.Context, *DeleteRoleRequest) (*emptypb.Empty, error)
	// EditRole编辑角色
	EditRole(context.Context, *EditRoleRequest) (*emptypb.Empty, error)
	// EditRoleDisabled 编辑角色状态
	EditRoleDisabled(context.Context, *EditRoleDisabledRequest) (*emptypb.Empty, error)
	// RoleDetail 角色详情
	RoleDetail(context.Context, *RoleDetailRequest) (*RoleDetailReply, error)
	// RoleList 角色列表
	RoleList(context.Context, *RoleListRequest) (*RoleListReply, error)
}

func RegisterRoleManageHTTPServer(s *http.Server, srv RoleManageHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/data/role/add", _RoleManage_AddRole0_HTTP_Handler(srv))
	r.POST("/v1/data/role/edit", _RoleManage_EditRole0_HTTP_Handler(srv))
	r.POST("/v1/data/role/edit/status", _RoleManage_EditRoleDisabled0_HTTP_Handler(srv))
	r.POST("/v1/data/role/delete", _RoleManage_DeleteRole0_HTTP_Handler(srv))
	r.GET("/v1/data/role/detail", _RoleManage_RoleDetail0_HTTP_Handler(srv))
	r.GET("/v1/data/role/list", _RoleManage_RoleList0_HTTP_Handler(srv))
}

func _RoleManage_AddRole0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageAddRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AddRole(ctx, req.(*AddRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddRoleReply)
		return ctx.Result(200, reply)
	}
}

func _RoleManage_EditRole0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageEditRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditRole(ctx, req.(*EditRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _RoleManage_EditRoleDisabled0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditRoleDisabledRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageEditRoleDisabled)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditRoleDisabled(ctx, req.(*EditRoleDisabledRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _RoleManage_DeleteRole0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageDeleteRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteRole(ctx, req.(*DeleteRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _RoleManage_RoleDetail0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RoleDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageRoleDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RoleDetail(ctx, req.(*RoleDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RoleDetailReply)
		return ctx.Result(200, reply)
	}
}

func _RoleManage_RoleList0_HTTP_Handler(srv RoleManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RoleListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleManageRoleList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RoleList(ctx, req.(*RoleListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RoleListReply)
		return ctx.Result(200, reply)
	}
}

type RoleManageHTTPClient interface {
	AddRole(ctx context.Context, req *AddRoleRequest, opts ...http.CallOption) (rsp *AddRoleReply, err error)
	DeleteRole(ctx context.Context, req *DeleteRoleRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditRole(ctx context.Context, req *EditRoleRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditRoleDisabled(ctx context.Context, req *EditRoleDisabledRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	RoleDetail(ctx context.Context, req *RoleDetailRequest, opts ...http.CallOption) (rsp *RoleDetailReply, err error)
	RoleList(ctx context.Context, req *RoleListRequest, opts ...http.CallOption) (rsp *RoleListReply, err error)
}

type RoleManageHTTPClientImpl struct {
	cc *http.Client
}

func NewRoleManageHTTPClient(client *http.Client) RoleManageHTTPClient {
	return &RoleManageHTTPClientImpl{client}
}

func (c *RoleManageHTTPClientImpl) AddRole(ctx context.Context, in *AddRoleRequest, opts ...http.CallOption) (*AddRoleReply, error) {
	var out AddRoleReply
	pattern := "/v1/data/role/add"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleManageAddRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *RoleManageHTTPClientImpl) DeleteRole(ctx context.Context, in *DeleteRoleRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/role/delete"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleManageDeleteRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *RoleManageHTTPClientImpl) EditRole(ctx context.Context, in *EditRoleRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/role/edit"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleManageEditRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *RoleManageHTTPClientImpl) EditRoleDisabled(ctx context.Context, in *EditRoleDisabledRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/role/edit/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleManageEditRoleDisabled))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *RoleManageHTTPClientImpl) RoleDetail(ctx context.Context, in *RoleDetailRequest, opts ...http.CallOption) (*RoleDetailReply, error) {
	var out RoleDetailReply
	pattern := "/v1/data/role/detail"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleManageRoleDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *RoleManageHTTPClientImpl) RoleList(ctx context.Context, in *RoleListRequest, opts ...http.CallOption) (*RoleListReply, error) {
	var out RoleListReply
	pattern := "/v1/data/role/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleManageRoleList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}