# data-mannagement/data-mannagement/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/data/role/add:
        post:
            tags:
                - RoleManage
            description: 新增角色
            operationId: RoleManage_AddRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rolemanage.v1.AddRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rolemanage.v1.AddRoleReply'
    /v1/data/role/delete:
        post:
            tags:
                - RoleManage
            description: 删除角色
            operationId: RoleManage_DeleteRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rolemanage.v1.DeleteRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/role/detail:
        get:
            tags:
                - RoleManage
            description: 角色详情
            operationId: RoleManage_RoleDetail
            parameters:
                - name: roleId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rolemanage.v1.RoleDetailReply'
    /v1/data/role/edit:
        post:
            tags:
                - RoleManage
            description: 编辑角色
            operationId: RoleManage_EditRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rolemanage.v1.EditRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/role/edit/status:
        post:
            tags:
                - RoleManage
            description: 编辑角色状态
            operationId: RoleManage_EditRoleDisabled
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rolemanage.v1.EditRoleDisabledRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/role/list:
        get:
            tags:
                - RoleManage
            description: 角色列表
            operationId: RoleManage_RoleList
            parameters:
                - name: pageSize
                  in: query
                  description: 每页条数（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: pageNumber
                  in: query
                  description: 页码数,从1开始（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: sortBy
                  in: query
                  description: 排序字段（非必须）
                  schema:
                    type: string
                - name: order
                  in: query
                  description: 排序方式("asc" or "desc") （非必须）
                  schema:
                    type: string
                - name: searchRoleName
                  in: query
                  description: 角色名称
                  schema:
                    type: string
                - name: searchRoleRemark
                  in: query
                  description: 角色备注
                  schema:
                    type: string
                - name: searchRoleDisabled
                  in: query
                  description: 按照是否禁用搜索, all：搜索全部，disabled：搜索禁用的 enabled：搜索未禁用的
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rolemanage.v1.RoleListReply'
...
components:
    schemas:
...
        api.rolemanage.v1.AddRoleReply:
            type: object
            properties:
                roleId:
                    type: integer
                    description: 角色id
                    format: uint32
        api.rolemanage.v1.AddRoleRequest:
            type: object
            properties:
                roleName:
                    type: string
                    description: 角色名称
                roleRemark:
                    type: string
                    description: 角色备注
        api.rolemanage.v1.DeleteRoleRequest:
            type: object
            properties:
                roleId:
                    type: integer
                    description: 角色id
                    format: uint32
        api.rolemanage.v1.EditRoleDisabledRequest:
            type: object
            properties:
                roleId:
                    type: integer
                    description: 角色id
                    format: uint32
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
        api.rolemanage.v1.EditRoleRequest:
            type: object
            properties:
                roleId:
                    type: integer
                    description: 用户id
                    format: uint32
                roleName:
                    type: string
                    description: 角色名称
                roleRemark:
                    type: string
                    description: 角色备注
        api.rolemanage.v1.RoleDetailReply:
            type: object
            properties:
                roleName:
                    type: string
                    description: 角色名称
                roleRemark:
                    type: string
                    description: 角色备注
                isDisabled:
                    type: boolean
                    description: 是否禁用
                isUsed:
                    type: boolean
                    description: 是否使用
                isBuiltin:
                    type: boolean
                    description: 是否内置
                updateTime:
                    type: integer
                    description: 更新时间
                    format: int64
                createTime:
                    type: integer
                    description: 生成时间
                    format: int64
        api.rolemanage.v1.RoleListReply:
            type: object
            properties:
                content:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rolemanage.v1.RoleListReply_Role'
                    description: 角色列表
                totalElements:
                    type: integer
                    description: 用户总个数
                    format: int32
        api.rolemanage.v1.RoleListReply_Role:
            type: object
            properties:
                roleName:
                    type: string
                    description: 角色名称
                roleRemark:
                    type: string
                    description: 角色备注
                isUsed:
                    type: boolean
                    description: 是否使用
                isBuiltin:
                    type: boolean
                    description: 是否内置
                isDisabled:
                    type: boolean
                    description: 是否禁用
                roleId:
                    type: integer
                    description: 角色id
                    format: uint32
                updateTime:
                    type: integer
                    description: 更新时间
                    format: int64
                createTime:
                    type: integer
                    description: 生成时间
                    format: int64
...